Wip
- configurable deployer-ecr bootstrap
- sync no-unused-vars eslint with tsconfig
- add unit test for deployer lambda
- refactor out lambda lib functions
