+++ infra/.eslintrc.json
@@ -12,7 +12,10 @@
     "node": true
   },
   "rules": {
-    "no-extra-semi": 2,
+    "no-extra-semi": "off",
+    "no-unused-vars": "off",
+    "@typescript-eslint/no-unused-vars": "error",
+    "@typescript-eslint/no-extra-semi": "error",
     "sort-imports": [
       "error",
       {
